Error catch for lastz
The fix I added checks alerts the user if lastz is not correctly install/broken on his or her computer. Previously, the code ran an operation checking the status of lztstderr, but didn't alert the user if an issue arose. Now, an else statement raises an exception which alerts the user if lastz is not working.