Updated working version of alignment trimming code
Following update to trim Velvet contigs of low coverage (lowercase)
bases, update alignment trimming code to filter slightly differently and
also add 2nd and 3rd pass filtering.

1) Changes to old code: now the code looks for alignment edges by moving across alignment
columns and requiring that (over a 20 bp sliding window) 65% of sites
have data for each taxon and 75% of those sites having data are
conserved across taxa.  The "edges" are found at either end in the first
windowed position where this is true.  Stuff outside the edges gets
trimmed.

2) Added 2nd pass filtering: after the first pass, above, the code
examines alignment data on a taxon-by-taxon basis (i.e., row-by-row  of
the alignment).  The code compares each row to the consensus for the
entire alignment and computes the running average (sliding window)
of differences between the taxon it's on and the consensus for all taxa.
The code then trims edges at the point where it finds running average
values < 0.5 prior to the first running average values of 0.99.  It does this on both ends, and converts
anything outside these edges to the insertion character "-".  This
should really be the missing data character "?", but it shouldn't really
matter much.  Changing to the missing data character to "?" should not
be a problem if we decide it's important.

3) Added 3rd pass filtering: because the conversion of edges in 2nd pass
filtering to "-" can basically reset the alignment boundaries (so all
taxa can sometimes start or end with a run of insertions like "----"),
re-run the first pass filtering again on the 2nd pass results to find
the edges of the alignment.