Update generic align to trim on 5 conserved bases at ends
The previous method was not behaving as well as I would like.  The
current method does the following during 2nd pass filtering:

After the first pass, the code examines alignment data on a
taxon-by-taxon basis (i.e., row-by-row  of the alignment).  The code
compares each row to the consensus for the entire alignment and computes
the running average (5 bp sliding window) of differences between the
taxon it's on and the consensus for all taxa.  The code then trims edges
at the point where it finds 5 adjacent running average values of >0.99.
It does this on both ends, and converts anything outside these edges to
the insertion character "-".  This should really be the missing data
character "?", but it shouldn't really matter much.  Changing to the
missing data character to "?" should not be a problem if we decide it's
important.