Update match_contigs_to_probes and fix fasta-finding code
Use CreateDir to ensure we always have a clean empty directory.  This means
we can remove the extra code to double-check this is true. Also update
fasta-finding code to deal better with different file types (and exclude
files that have an extension like *.fa* but are not fasta files (e.g. *.fai
files).